# backend/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkyCrop API
  version: 0.1.0
  description: |
    SkyCrop - Satellite-Based Paddy Field Management & Monitoring System
    MVP Auth and Health endpoints.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.skycrop.com
    description: Production server
...
paths:
...
  /api/v1/fields:
    get:
      tags: [Fields]
      summary: List user's fields (with spatial filters)
      description: Returns active fields owned by the authenticated user with optional spatial filters and pagination.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: bbox
          schema:
            type: string
            pattern: '^-?\\d{1,3}\\.\\d+,-?\\d{1,2}\\.\\d+,-?\\d{1,3}\\.\\d+,-?\\d{1,2}\\.\\d+$'
          description: 'minLon,minLat,maxLon,maxLat (SRID 4326)'
        - in: query
          name: near
          schema:
            type: string
            pattern: '^-?\\d{1,2}\\.\\d+,-?\\d{1,3}\\.\\d+,\\d{1,6}$'
          description: 'lat,lon,radius_m (uses ST_DWithin on center)'
        - in: query
          name: intersects
          schema:
            type: string
          description: 'GeoJSON Polygon/MultiPolygon as string or "field:{uuid}"'
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: sort
          schema: { type: string, enum: [name, created_at, area_sqm], default: created_at }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: List of fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Field'
                  pagination:
                    type: object
                    properties:
                      page: { type: integer, example: 1 }
                      page_size: { type: integer, example: 20 }
                      total: { type: integer, example: 2 }
                required: [success, data]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Fields]
      summary: Create a new field (manual boundary)
      description: Creates a field using a provided boundary polygon and name
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldCreateRequest'
      responses:
        '201':
          description: Field created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Field'
                  warnings:
                    type: array
                    items:
                      type: object
                      properties:
                        code: { type: string, example: GEOMETRY_SNAPPED }
                        message: { type: string }
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/fields/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Fields]
      summary: Get field by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Field details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                required: [success, data]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Fields]
      summary: Update field metadata (name/status)
      description: Update field name
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # required removed; partial update (name/status) per Sprint 2
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                status:
                  type: string
                  enum: [active, archived, deleted]
      responses:
        '200':
          description: Field updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Fields]
      summary: Delete field (soft delete)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Field deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/ml/segmentation/predict:
    post:
      tags: [ML]
      summary: Segmentation mask prediction
      description: Produces segmentation mask for bbox or field_id. Proxies to internal ML service.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MLPredictRequest'
      responses:
        '200':
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MLMaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '504':
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          example: null
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Email is invalid
            details:
              type: object
              nullable: true
      required: [success, error]
...
    GeoJSONPolygon:
      type: object
      description: GeoJSON Polygon or MultiPolygon in SRID 4326
      oneOf:
        - type: object
          properties:
            type: { type: string, enum: [Polygon] }
            coordinates:
              type: array
              description: Array of linear rings (first is outer ring, subsequent are holes)
              items:
                type: array
                items:
                  type: array
                  minItems: 2
                  items:
                    type: number
          required: [type, coordinates]
        - type: object
          properties:
            type: { type: string, enum: [MultiPolygon] }
            coordinates:
              type: array
              items:
                type: array
    Field:
      type: object
      properties:
        field_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        boundary:
          $ref: '#/components/schemas/GeoJSONPolygon'
        area_sqm:
          type: number
          format: float
          description: Area in square meters computed via ST_Area(geography)
        center:
          type: object
          description: GeoJSON Point
          properties:
            type: { type: string, enum: [Point] }
            coordinates:
              type: array
              minItems: 2
              maxItems: 3
              items: { type: number }
        status: { type: string, enum: [active, archived, deleted] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required:
        [field_id, user_id, name, boundary, area_sqm, center, status, created_at, updated_at]
    FieldCreateRequest:
      type: object
      required: [name, boundary]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        boundary:
          $ref: '#/components/schemas/GeoJSONPolygon'
    MLPredictRequest:
      type: object
      properties:
        bbox:
          type: array
          items: { type: number }
          minItems: 4
          maxItems: 4
          description: [minLon, minLat, maxLon, maxLat]
        field_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        model_version:
          type: string
          example: unet-1.0.0
        tiling:
          type: object
          properties:
            size: { type: integer, default: 512 }
            overlap: { type: integer, default: 64 }
        return:
          type: string
          enum: [mask_url, inline]
          default: mask_url
      oneOf:
        - required: [bbox, date]
        - required: [field_id, date]
    MLMaskResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            request_id: { type: string }
            model:
              type: object
              properties:
                name: { type: string, example: unet }
                version: { type: string, example: 1.0.0 }
            mask_url: { type: string, format: uri }
            mask_base64: { type: string, description: inline encoded mask when requested }
            mask_format: { type: string, enum: [geojson, png] }
            metrics:
              type: object
              properties:
                latency_ms: { type: integer }
                tile_count: { type: integer }
                cloud_coverage: { type: number, format: float }
            warnings:
              type: array
              items: { type: string }
...